fix: Split out client and server transport code into separate modules and also split out the http client code into a separate module
