+++ client-http/src/main/java/io/a2a/client/http/A2AHttpResponse.java
@@ -1,0 +1,9 @@
+package io.a2a.client.http;
+
+public interface A2AHttpResponse {
+    int status();
+
+    boolean success();
+
+    String body();
+}
