+++ client/src/test/java/io/a2a/client/A2ACardResolverTest.java
@@ -11,8 +11,8 @@ import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import io.a2a.transport.jsonrpc.client.A2AHttpClient;
-import io.a2a.transport.jsonrpc.client.A2AHttpResponse;
+import io.a2a.client.http.A2AHttpClient;
+import io.a2a.client.http.A2AHttpResponse;
 import io.a2a.spec.A2AClientError;
 import io.a2a.spec.A2AClientJSONError;
 import io.a2a.spec.AgentCard;
