+++ pom.xml
@@ -280,6 +280,10 @@
 
     <modules>
         <module>client</module>
+        <module>client-http</module>
+        <module>client-transport/grpc</module>
+        <module>client-transport/jsonrpc</module>
+        <module>client-transport/spi</module>
         <module>common</module>
         <module>examples/helloworld</module>
         <module>reference/common</module>
@@ -290,7 +294,6 @@
         <module>spec-grpc</module>
         <module>tck</module>
         <module>tests/server-common</module>
-        <module>transport/spi</module>
         <module>transport/jsonrpc</module>
         <module>transport/grpc</module>
     </modules>
