+++ server-common/pom.xml
@@ -24,7 +24,7 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>a2a-java-sdk-client</artifactId>
+            <artifactId>a2a-java-sdk-client-http</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
