--- server-common/src/main/java/io/a2a/server/requesthandlers/CallContextFactory.java
@@ -1,8 +1,0 @@
-package io.a2a.server.requesthandlers;
-
-import io.a2a.server.ServerCallContext;
-import io.grpc.stub.StreamObserver;
-
-public interface CallContextFactory {
-    <V> ServerCallContext create(StreamObserver<V> responseObserver);
-}
