--- transport/jsonrpc/src/main/java/io/a2a/transport/jsonrpc/client/A2AHttpResponse.java
@@ -1,9 +1,0 @@
-package io.a2a.transport.jsonrpc.client;
-
-public interface A2AHttpResponse {
-    int status();
-
-    boolean success();
-
-    String body();
-}
